fix: delete endpoint on push notification disable
fix: add endpoint deletion on disable

fix: use definemessages util

refactor: add code comment

+++ server/routes/user/index.ts
@@ -269,15 +269,19 @@ router.delete<{ userId: number; endpoint: string }>(
     try {
       const userPushSubRepository = getRepository(UserPushSubscription);
 
-      const userPushSub = await userPushSubRepository.findOneOrFail({
-        relations: {
-          user: true,
-        },
+      const userPushSub = await userPushSubRepository.findOne({
+        relations: { user: true },
         where: {
           user: { id: req.params.userId },
           endpoint: req.params.endpoint,
         },
       });
+
+      // If not found, just return 204 to prevent push disable failure
+      // (rare scenario where user push sub does not exist)
+      if (!userPushSub) {
+        return res.status(204).send();
+      }
 
       await userPushSubRepository.remove(userPushSub);
       return res.status(204).send();
